changed props to state
