+++ src/components/PostForm/PostForm.js
@@ -1,6 +1,8 @@
-import {useState, useRef, useEffect } from 'react'
+import {useState, useRef, useEffect, useCallback } from 'react'
 import { makeStyles, TextField, Paper, Button, Icon, Grid } from '@material-ui/core'
 import PropTypes from 'prop-types'
+import { useDispatch } from 'react-redux'
+import { onAddMessageAction } from '../../store/actions'
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -17,8 +19,18 @@ const useStyles = makeStyles((theme) => ({
   }
 }));
 
-function PostForm ({onAdd=f=>f, formRef}) {
+function PostForm ({chatId, formRef}) {
 
+  const dispatch = useDispatch()
+  
+  const onAdd = useCallback(
+    (message) => {
+      message = 
+      dispatch(onAddMessageAction(message, chatId))
+    },
+    [dispatch, chatId],
+  )
+  
   const classes = useStyles()
 
   const timer = useRef(null)
@@ -136,7 +148,7 @@ function PostForm ({onAdd=f=>f, formRef}) {
 }
 
 PostForm.propTypes = {
-  onAdd: PropTypes.func,
+  chatId: PropTypes.string,
   formRef: PropTypes.object
 }
